+++ classes/paginationpage.php
@@ -1,12 +1,10 @@
 <?php
 namespace Grav\Plugin;
 
-use \Grav\Common\GravTrait;
+use \Grav\Common\Grav;
 
 class PaginationPage
 {
-    use GravTrait;
-
     /**
      * @var int
      */
@@ -32,7 +30,7 @@ class PaginationPage
     {
         $this->number = $number;
         $this->url = $url;
-        $this->delta = self::getGrav()['config']->get('plugins.pagination.delta');
+        $this->delta = Grav::instance()['config']->get('plugins.pagination.delta');
     }
 
     /**
@@ -42,7 +40,7 @@ class PaginationPage
      */
     public function isCurrent()
     {
-        if (self::getGrav()['uri']->currentPage() == $this->number) {
+        if (Grav::instance()['uri']->currentPage() == $this->number) {
             return true;
         } else {
             return false;
@@ -59,7 +57,7 @@ class PaginationPage
         if (!$this->delta) {
             return true;
         } else {
-            return abs(self::getGrav()['uri']->currentPage() - $this->number) < $this->delta;
+            return abs(Grav::instance()['uri']->currentPage() - $this->number) < $this->delta;
         }
     }
 
@@ -74,7 +72,7 @@ class PaginationPage
         if (!$this->delta) {
             return false;
         } else {
-            return abs(self::getGrav()['uri']->currentPage() - $this->number) == $this->delta;
+            return abs(Grav::instance()['uri']->currentPage() - $this->number) == $this->delta;
         }
     }
 }